Remove option to provide ApplicationContextInitializer.
The previous change to be able to provide both @Configuration classes
or XML config locations limits what can be done with an
ApplicationContextInitializer because we no longer do all the work
through an AnnotationConfig- or an Xml- WebApplicationContext and
instead provide a populated BeanFactory to the constructor of a
GenericWebApplicationContext. As a result an
ApplicationContextInitializer can't specify config locations,
active profiles, etc.

Instead we should consider providing an additional
MockMvcBuilders factory method that accepts a
WebApplicationContextInitializer and the type of
ConfiguratbleWebApplicationContext class to instantiate.